amazon-eks/cdk/requirements.txt
-e .
aws-cdk.core==1.140.0
aws-cdk.aws-eks==1.140.0
aws-cdk.aws-iam==1.140.0
aws-cdk.aws-ec2==1.140.0
aws-cdk.aws-kms==1.140.0

